- added express-validator - [bonus task 2] all requests should be properly validated, using express-validator where possible, and basic validation where not

+++ src/controllers/trackingController.ts
@@ -1,12 +1,15 @@
 import {NextFunction, Request, Response} from "express"
 import {UserModel} from "../db/user";
 import {models} from "../db";
+import {validationResult} from "express-validator";
 
 const {
 	Exercise,
 	ExerciseHistory
 } = models
 
+//	GET
+//	Returns all tracked exercises of logged-in user
 exports.getTrackedExercises = async (_req: Request, res: Response, _next: NextFunction) => {
 	const user = _req.user as UserModel
 
@@ -26,8 +29,16 @@ exports.getTrackedExercises = async (_req: Request, res: Response, _next: NextFu
 	})
 }
 
+//	POST
+//	Creates tracked exercise entry
+//	Requires ID of tracked exercise, other fields can be left empty if desired
 exports.postTrackedExercise = async (_req: Request, res: Response, _next: NextFunction) => {
 	const user = _req.user as UserModel
+
+	const validRes = validationResult(_req);
+	if (!validRes.isEmpty()) {
+		return res.status(400).send('Invalid request parameters')
+	}
 
 	try {
 		const createdExercise = await ExerciseHistory.create({
@@ -47,8 +58,16 @@ exports.postTrackedExercise = async (_req: Request, res: Response, _next: NextFu
 	}
 }
 
+//	DELETE
+//	Removes tracked exercise based on ID
+//	Note: user can only remove their own tracked exercise
 exports.removeTrackedExercise = async (_req: Request, res: Response, _next: NextFunction) => {
 	const user = _req.user as UserModel
+
+	const validRes = validationResult(_req);
+	if (!validRes.isEmpty()) {
+		return res.status(400).send('Invalid request parameters')
+	}
 
 	try {
 		const rowsUpdated = await ExerciseHistory.destroy({
